+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Crash on API 10 devices (#3726).
 * `UnsatisfiedLinkError` caused by `pipe2` (#3945).
 * Unrecoverable error with message "Try again" when the notification fifo is full (#3964).
+* Realm migration wasn't triggered when the primary key definition was altered (#3966).
 
 ### Enhancements